refactor: reduce redundant test files
